Base for quality is now a command line switch. Improvement (?) to command line parsing.
git-svn-id: https://svn.cs.umd.edu/repos/carlksvn/jellyfish/src@1005 3840e3e0-9109-4a78-8b6e-c2447a23df0a

+++ tests/parallel_direct_indexing.sh
@@ -1,10 +1,14 @@
 #! /bin/sh
 
-nCPUs=$(grep -c '^processor' /proc/cpuinfo)
+nCPUs=$(grep -c '^processor' /proc/cpuinfo 2>/dev/null || sysctl -n hw.ncpu)
 pref=$(basename $0 .sh)
-echo "Counting 10-mers on ${nCPUs} CPU"
-../jellyfish/jellyfish count --matrix seq10m_matrix_10 -m 10 -t $nCPUs -o $pref -s 10000000 seq10m.fa && \
+JF=../jellyfish/jellyfish
+echo "Counting 10-mers on ${nCPUs} CPU" && \
+    $JF count --matrix seq10m_matrix_10 -m 10 -t $nCPUs \
+    -o $pref -s 10000000 --timing ${pref}.timing seq10m.fa && \
     echo "dcbb23c4a74a923c37a3b059f6a6d89a  ${pref}_0" | md5sum -c
 RET=$?
-rm -f ${pref}_*
+
+cat ${pref}.timing
+[ -z "$NODEL" ] && rm -f ${pref}_* ${pref}.timing
 exit $RET
